2740 projekty optimalizace rychlosti načítání mapy (#2880)
* Odstranění nepoužitých funkcí

* Optimalizace načítání dat z databáze

* Optimalizace načítání projektů do mapy

* Bugfix circular import

* Vrácení změn

* Oprava kruhového importu

---------

+++ webclient/xml_generator/models.py
@@ -72,8 +72,6 @@ class ModelWithMetadata(BaseAmcrModel):
         self.deletion_record_saved = False
         self.skip_container_check = False
         super(ModelWithMetadata, self).__init__(*args, **kwargs)
-        if not hasattr(self, "soubory"):
-            self.soubory = None
 
     def create_transaction(self, transaction_user, success_message=None, error_message=None):
         from core.repository_connector import FedoraTransaction
@@ -212,13 +210,13 @@ class ModelWithMetadata(BaseAmcrModel):
             from pas.models import SamostatnyNalez
 
             if isinstance(self, Dokument):
-                if self.soubory.pk is not None:
+                if hasattr(self, "soubory") and self.soubory.pk is not None:
                     for item in self.soubory.soubory.all():
                         item.suppress_signal = True
                         item.delete()
                     self.soubory.delete()
             elif isinstance(self, SamostatnyNalez):
-                if self.soubory.pk is not None:
+                if hasattr(self, "soubory") and self.soubory.pk is not None:
                     for item in self.soubory.soubory.all():
                         item.suppress_signal = True
                         item.delete()
